refactor(semantic): consolidate CFG reachability traversal
